# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Realona Exchange API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get all active listings
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: game
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: divisionRank
          in: query
          schema:
            type: string
        - name: minSquadRating
          in: query
          schema:
            type: integer
        - name: maxSquadRating
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
...
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a listing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── TRADES ────────────────────────────────────────────────────────────────
...
  /listings/{id}/offers:
    post:
      operationId: makeOffer
      tags: [offers]
      summary: Make an offer on a listing
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OfferInput" }
      responses:
        "201":
          description: Offer created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Offer" }
...
  /offers/{id}/respond:
    post:
      operationId: respondToOffer
      tags: [offers]
      summary: Accept, reject, or counter an offer
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OfferResponse" }
      responses:
        "200":
          description: Offer updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessResponse" }

  # ─── KYC ───────────────────────────────────────────────────────────────────
  /kyc/submit:
    post:
      operationId: submitKyc
      tags: [kyc]
      summary: Submit KYC documents for verification
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KycInput" }
      responses:
        "201":
          description: KYC submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KycSubmission" }
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ── Auth ──
...
    Listing:
      type: object
      required: [id, sellerId, gameName, price, description, status, category, createdAt]
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerUsername:
          type: ["string", "null"]
        sellerIsVerified:
          type: boolean
        sellerRating:
          type: ["number", "null"]
        category:
          type: string
          enum: [efootball, social_media]
        gameName:
          type: string
        price:
          type: number
        description:
          type: string
        pictureUrl:
          type: ["string", "null"]
        accountEmail:
          type: ["string", "null"]
        accountPassword:
          type: ["string", "null"]
        divisionRank:
          type: ["string", "null"]
        squadRating:
          type: ["integer", "null"]
        # social media fields
        platform:
          type: ["string", "null"]
        accountHandle:
          type: ["string", "null"]
        followerCount:
          type: ["integer", "null"]
        following:
          type: ["integer", "null"]
        accountAge:
          type: ["string", "null"]
        engagementRate:
          type: ["string", "null"]
        viewCount:
          type: integer
        status:
          type: string
          enum: [active, sold, deleted]
        createdAt:
          type: string
          format: date-time
    ListingInput:
      type: object
      required: [gameName, price, description]
      properties:
        category:
          type: string
          enum: [efootball, social_media]
        gameName:
          type: string
        price:
          type: number
        description:
          type: string
        pictureUrl:
          type: string
        accountEmail:
          type: string
        accountPassword:
          type: string
        konamiId:
          type: string
        konamiPassword:
          type: string
        accessCode:
          type: string
        divisionRank:
          type: string
        squadRating:
          type: integer
          minimum: 2000
          maximum: 5000
        platform:
          type: string
        accountHandle:
          type: string
        followerCount:
          type: integer
        following:
          type: integer
        accountAge:
          type: string
        engagementRate:
          type: string
    ListingUpdate:
      type: object
      properties:
        gameName:
          type: string
        price:
          type: number
        description:
          type: string
        pictureUrl:
          type: string
        status:
          type: string

    # ── Trade ──
...
    Offer:
      type: object
      required: [id, listingId, buyerId, sellerId, amount, status, expiresAt, createdAt]
      properties:
        id:
          type: integer
        listingId:
          type: integer
        buyerId:
          type: integer
        sellerId:
          type: integer
        buyerUsername:
          type: ["string", "null"]
        sellerUsername:
          type: ["string", "null"]
        listingTitle:
          type: ["string", "null"]
        amount:
          type: number
        counterAmount:
          type: ["number", "null"]
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected, countered, expired]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    OfferInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        message:
          type: string
    OfferResponse:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accept, reject, counter]
        counterAmount:
          type: number
    KycSubmission:
      type: object
      required: [id, userId, documentType, documentUrl, status, level, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: ["string", "null"]
        documentType:
          type: string
        documentUrl:
          type: string
        selfieUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        level:
          type: integer
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    KycInput:
      type: object
      required: [documentType, documentUrl]
      properties:
        documentType:
          type: string
        documentUrl:
          type: string
        selfieUrl:
          type: string
...